Convert storage service version to a long.

+++ idp-authn-api/src/main/java/net/shibboleth/idp/authn/AuthenticationFlowDescriptor.java
@@ -299,7 +299,7 @@ public class AuthenticationFlowDescriptor extends AbstractIdentifiableInitializa
     }
 
     /** {@inheritDoc} */
-    @Override @Nonnull public AuthenticationResult deserialize(int version, @Nonnull @NotEmpty final String context,
+    @Override @Nonnull public AuthenticationResult deserialize(long version, @Nonnull @NotEmpty final String context,
             @Nonnull @NotEmpty final String key, @Nonnull @NotEmpty final String value, @Nonnull final Long expiration)
             throws IOException {
         ComponentSupport.ifNotInitializedThrowUninitializedComponentException(this);
